Fixed Red-5 + Refactor Rename
