+++ src/engine/pieces/piece.ts
@@ -19,8 +19,57 @@ export default class Piece {
         board.movePiece(currentSquare, newSquare);
     }
 
-    protected inBounds(position: Square) {
+    private inBounds(position: Square) {
         return (position.row >= 0 && position.row < GameSettings.BOARD_SIZE &&
                 position.col >= 0 && position.col < GameSettings.BOARD_SIZE);
+    }
+
+    protected getRookMoves(currentPosition: Square, direction: string): Square[] {
+        if (direction !== 'horizontal' && direction !== 'vertical') {
+            throw new Error('Unexpected direction.');
+        }
+
+        const constantDirection = (direction === 'horizontal') ? currentPosition.row : currentPosition.col;
+        const otherCoordinate = (direction === 'horizontal') ? currentPosition.col : currentPosition.row;
+
+        let result: Square[] = [] as Square[];
+        for (let i: number = 0; i < GameSettings.BOARD_SIZE; i += 1) {
+            if (i === otherCoordinate) {
+                continue;
+            }
+
+            const newSquare: Square = (direction === 'horizontal') ? new Square(constantDirection, i) : new Square(i, constantDirection);
+            result.push(newSquare);
+        }
+
+        return result;
+    }
+
+    protected getBishopMoves(currentPosition: Square, direction: string): Square[] {
+        if (direction !== 'main' && direction !== 'secondary') {
+            throw new Error('Unexpected diagonal.');
+        }
+
+        let row: number = currentPosition.row - 1;
+        let col: number = currentPosition.col + (direction === 'main' ? -1 : 1);
+
+        let result: Square[] = [] as Square[];
+        while (this.inBounds(new Square(row, col))) {
+            result.push(new Square(row, col));
+
+            row -= 1;
+            col += (direction === 'main' ? -1 : 1);
+        }
+
+        row = currentPosition.row + 1;
+        col = currentPosition.col + (direction === 'main' ? 1 : -1);
+        while (this.inBounds(new Square(row, col))) {
+            result.push(new Square(row, col));
+
+            row += 1;
+            col += (direction === 'main' ? 1 : -1);
+        }
+
+        return result;
     }
 }
