+++ src/engine/pieces/rook.ts
@@ -2,7 +2,6 @@ import Piece from './piece';
 import Player from '../player';
 import Board from '../board';
 import Square from '../square';
-import GameSettings from '../gameSettings';
 
 export default class Rook extends Piece {
     public constructor(player: Player) {
@@ -11,27 +10,6 @@ export default class Rook extends Piece {
 
     public getAvailableMoves(board: Board) {
         const currentPosition: Square = board.findPiece(this);
-        return this.getMovesPerDirection(currentPosition, 'horizontal').concat(this.getMovesPerDirection(currentPosition, 'vertical'));
-    }
-
-    private getMovesPerDirection(currentPosition: Square, direction: string): Square[] {
-        if (direction !== 'horizontal' && direction !== 'vertical') {
-            throw new Error('Unexpected direction.');
-        }
-
-        const constantDirection = (direction === 'horizontal') ? currentPosition.row : currentPosition.col;
-        const otherCoordinate = (direction === 'horizontal') ? currentPosition.col : currentPosition.row;
-
-        let result: Square[] = [] as Square[];
-        for (let i: number = 0; i < GameSettings.BOARD_SIZE; i += 1) {
-            if (i === otherCoordinate) {
-                continue;
-            }
-
-            const newSquare: Square = (direction === 'horizontal') ? new Square(constantDirection, i) : new Square(i, constantDirection);
-            result.push(newSquare);
-        }
-
-        return result;
+        return this.getRookMoves(currentPosition, 'horizontal').concat(this.getRookMoves(currentPosition, 'vertical'));
     }
 }
